+++ src/react/src/components/PartnerFields/FormatComponents/UriReferenceProperty/UriReferenceProperty.jsx
@@ -2,6 +2,7 @@ import React from 'react';
 import { string, object, shape } from 'prop-types';
 import { withStyles } from '@material-ui/core/styles';
 import Typography from '@material-ui/core/Typography';
+import OpenInNewIcon from '@material-ui/icons/OpenInNew';
 import { getDescription, getAbsoluteUri, getDisplayLinkText } from './utils';
 import { getLinkTextFromSchema } from '../../utils';
 import { commonPropertyStyles } from '../../styles';
@@ -53,8 +54,10 @@ const UriReferenceProperty = ({
                 href={absoluteUri}
                 target="_blank"
                 rel="noopener noreferrer"
+                className={classes.link}
             >
-                {displayLinkText}
+                <span>{displayLinkText}</span>
+                <OpenInNewIcon className={classes.linkIcon} />
             </a>
         </div>
     );