+++ src/react/src/components/PartnerFields/PartnerFieldLabel/PartnerFieldLabel.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { string } from 'prop-types';
+import { object, string } from 'prop-types';
 import { withStyles } from '@material-ui/core/styles';
 import partnerFieldLabelStyles from './styles';
 
@@ -9,6 +9,7 @@ const PartnerFieldLabel = ({ title, classes }) => (
 
 PartnerFieldLabel.propTypes = {
     title: string.isRequired,
+    classes: object.isRequired,
 };
 
 export default withStyles(partnerFieldLabelStyles)(PartnerFieldLabel);