+++ src/react/src/components/PartnerFields/styles.js
@@ -2,6 +2,15 @@ export const commonPropertyStyles = {
     container: {
         marginBottom: '4px',
     },
+    link: {
+        display: 'inline-flex',
+        alignItems: 'center',
+    },
+    linkIcon: {
+        marginTop: 6,
+        fontSize: 12,
+        marginLeft: 2,
+    },
 };
 
 export default commonPropertyStyles;